+++ groupShopper/AddPeoplePage.swift
@@ -8,8 +8,7 @@
 
 import UIKit
 
-class ViewController: UIViewController, UITableViewDelegate, UITableViewDataSource {
-
+class AddPeople: UIViewController, UITableViewDelegate, UITableViewDataSource {
     
     var tableView = UITableView(frame: CGRect(), style: .plain)
     var button = UIButton(frame: CGRect(x: 0, y: 0, width: 100, height: 30))
@@ -18,37 +17,41 @@ class ViewController: UIViewController, UITableViewDelegate, UITableViewDataSour
         super.viewDidLoad()
         
         title = "Add People"
+        view.backgroundColor = UIColor.white
         
-        personList = ["Apple", "Banana", "Carrot"]
+        personList = ["Apple", "Banana", "Carrot"] // temp, will be populated in this class
         
-        button.setTitle("Press this", for: .normal)
+        
+        // Button to advance to next page
+        button.setTitle("Done", for: .normal)
         button.setTitleColor(UIColor.blue, for: .normal)
-        button.addTarget(self, action: #selector(ViewController.showNextScreen), for: .touchUpInside)
+        button.addTarget(self, action: #selector(AddPeople.showNextScreen), for: .touchUpInside)
         
+        view.addSubview(button)
+        
+        // Displays people in a list on page
         tableView.dataSource = self
         tableView.delegate = self
         tableView.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
         
         view.addSubview(tableView)
-        view.addSubview(button)
         
-        view.backgroundColor = UIColor.white
     }
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         
-        button.frame = CGRect(x: 0, y: view.frame.height - 50, width: view.frame.width/2, height: 50)
+        // Set advance button frame to be placed right justafied on the bottom of the screen // COULD USE CLEANUP ON PLACEMENT
+        button.frame = CGRect(x: view.frame.width - 150, y: view.frame.height - 50, width: view.frame.width / 2, height: 50)
         button.reloadInputViews()
         
+        // List frame takes up all but bottom of screen to allow room for "done" button
         tableView.frame = CGRect(x: 0, y: 0, width: view.frame.width, height: view.frame.height - 50)
         tableView.reloadData()
     }
     
     
     @objc func showNextScreen() {
-        //let vc = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "AddItemsPage") as! AddItemsPage
-        //navigationController?.pushViewController(vc, animated: true)
         navigationController?.pushViewController(AddItemsPage(), animated: true)
     }
     
@@ -58,9 +61,7 @@ class ViewController: UIViewController, UITableViewDelegate, UITableViewDataSour
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
-        
         cell.textLabel?.text = personList[indexPath.row]
-        
         return cell
     }
